fix module deleting when corresponding service is inactive
This shot on wbio-ai-dv-12 when wb-mqtt-adc service is inactive.

Problem is in `[[ x ]] && { y }` statement: when `[[ x ]]` is false,
bash returns false for the whole sentence.

Because of this and `while` in the wb-hwconf-helper, delete procedure
entered an infinite loop.

`if [[ x ]]; then y; fi` statement is better because does not return
bad exit codes, so, no loops here.